Moves the state up for the map components to avoid bubbling issues
