+++ src/pages/monitoring/outlets/initiativesMap/MapFinder.tsx
@@ -8,14 +8,10 @@ import type {
   FeatureCollection,
   GeoJsonProperties,
   Geometry,
-  MultiPolygon,
-  Polygon,
 } from "geojson";
 
-import { ErrorsList } from "@ui/LabelingWithErrors";
 import { INITIAVIVES_MAP_GRADIENT, COUNTRY_BOUNDS } from "@config/monitoring";
 
-import { isMonitoringAPIError } from "pages/monitoring/api/types/guards";
 import { createGradientScale } from "pages/monitoring/utils/createGradientScale";
 import { type InitiativeByLocation } from "pages/monitoring/types/initiative";
 import { ChangeView } from "pages/monitoring/outlets/initiativesMap/mapFinder/ChangeView";
@@ -24,54 +20,39 @@ import {
   MapMarker,
 } from "pages/monitoring/outlets/initiativesMap/mapFinder/MapMarker";
 import { ZoomControls } from "pages/monitoring/outlets/initiativesMap/mapFinder/ZoomControls";
-import { MapLegend } from "pages/monitoring/outlets/initiativesMap/mapFinder/MapLegend";
 import {
   type MAP_LAYERS,
   MAP_TILES,
 } from "pages/monitoring/outlets/initiativesMap/layout/layers";
-import { getGeoJsonMap } from "pages/monitoring/api/services/location";
-import { getInitiativeLocations } from "pages/monitoring/api/services/initiatives";
+import type {
+  DeptFeature,
+  DeptProperties,
+} from "pages/monitoring/outlets/initiativesMap/types/mapFeatures";
 
-interface DeptProperties {
-  geofence_name: string;
-  gid: number;
-}
-
-type DeptFeature = Feature<Polygon | MultiPolygon, DeptProperties>;
-
-export function MapFinder() {
+export function MapFinder({
+  tiles,
+  layer,
+  nation,
+  initiatives,
+  activeFeatures,
+  leastInitiativesPerDepartment,
+  mostInitiativesPerDepartment,
+}: {
+  tiles: keyof typeof MAP_TILES;
+  layer: keyof typeof MAP_LAYERS | null;
+  nation: FeatureCollection | null;
+  initiatives: InitiativeByLocation[];
+  activeFeatures: {
+    feature: Feature<Geometry, GeoJsonProperties>;
+    count: number;
+  }[];
+  leastInitiativesPerDepartment: number;
+  mostInitiativesPerDepartment: number;
+}) {
   const { departmentId, initiativeId } = useParams();
   const navigate = useNavigate();
-
-  const [errors, setErrors] = useState<string[]>([]);
   const [center, setCenter] = useState<L.LatLng | null>(null);
   const [bounds, setBounds] = useState<LatLngBoundsLiteral | null>(null);
-  const [nation, setNation] = useState<FeatureCollection | null>(null);
-  const [initiatives, setInitiatives] = useState<InitiativeByLocation[]>([]);
-  const [tiles, setTiles] = useState<keyof typeof MAP_TILES>(0);
-  const [layer, setLayer] = useState<keyof typeof MAP_LAYERS | null>(null);
-
-  useEffect(() => {
-    const fetchMapInfo = async () => {
-      const mapInfo = await getGeoJsonMap();
-
-      if (isMonitoringAPIError(mapInfo)) {
-        setErrors(mapInfo.data.map((err) => err.msg));
-        return;
-      }
-      setNation(mapInfo);
-
-      const initiativesLocations = await getInitiativeLocations();
-
-      if (isMonitoringAPIError(initiativesLocations)) {
-        setInitiatives([]);
-        return;
-      }
-      setInitiatives(initiativesLocations);
-    };
-
-    void fetchMapInfo();
-  }, []);
 
   useEffect(() => {
     if (!nation || !nation.features) {
@@ -115,84 +96,19 @@ export function MapFinder() {
     }
   }, [departmentId, nation, initiativeId, initiatives]);
 
-  const processedData = useMemo<
-    { feature: Feature<Geometry, GeoJsonProperties>; count: number }[]
-  >(() => {
-    if (!initiatives.length || !nation || !nation?.features) {
-      return [];
-    }
-
-    const groupedInitiatives = initiatives.reduce<Record<number, number>>(
-      (all, current) => {
-        if (!all[current.mainLocationId]) {
-          all[current.mainLocationId] = 0;
-        }
-        all[current.mainLocationId] += 1;
-        return all;
-      },
-      {},
-    );
-
-    const results = [];
-
-    for (const feature of nation.features as Feature<
-      Polygon | MultiPolygon
-    >[]) {
-      if (groupedInitiatives[feature.properties?.gid as number] > 0) {
-        results.push({
-          feature,
-          count: groupedInitiatives[feature.properties?.gid as number],
-        });
-      }
-    }
-
-    return results;
-  }, [initiatives, nation]);
-
-  const departmentsWithInitiatives = useMemo(
-    () =>
-      processedData.map(({ feature }) => {
-        const f = feature.properties as DeptProperties;
-        return { value: String(f.gid), label: f.geofence_name };
-      }),
-    [processedData],
-  );
-
-  const [min, max] = useMemo(() => {
-    let [currentMin, currentMax] = processedData.reduce(
-      (result, current) => {
-        if (current.count <= result[0]) {
-          result[0] = current.count;
-        }
-        if (current.count >= result[1]) {
-          result[1] = current.count;
-        }
-        return result;
-      },
-      [Infinity, 0],
-    );
-
-    if (currentMin === Infinity) {
-      currentMin = 1;
-    }
-    if (currentMax === 0) {
-      currentMax = 2;
-    }
-    if (currentMin === currentMax) {
-      currentMax++;
-    }
-
-    return [currentMin, currentMax];
-  }, [processedData]);
-
   const getColor = useMemo(
-    () => createGradientScale(min, max, INITIAVIVES_MAP_GRADIENT),
-    [min, max],
+    () =>
+      createGradientScale(
+        leastInitiativesPerDepartment,
+        mostInitiativesPerDepartment,
+        INITIAVIVES_MAP_GRADIENT,
+      ),
+    [leastInitiativesPerDepartment, mostInitiativesPerDepartment],
   );
 
   const setDeptStyle = (feature?: Feature) => {
     const f = feature as DeptFeature;
-    const dataItem = processedData.find(
+    const dataItem = activeFeatures.find(
       (d) =>
         d.feature.properties?.geofence_name === f.properties?.geofence_name,
     );
@@ -212,7 +128,7 @@ export function MapFinder() {
   const setFeatureBehavior = (feature: Feature, layer: L.Layer) => {
     const f = feature as DeptFeature;
 
-    const dataItem = processedData.find(
+    const dataItem = activeFeatures.find(
       (d) =>
         d.feature.properties?.geofence_name === f.properties?.geofence_name,
     );
@@ -234,67 +150,50 @@ export function MapFinder() {
     });
   };
 
-  return errors.length > 0 ? (
-    <ErrorsList errorItems={errors} />
-  ) : (
-    <>
-      <MapContainer
-        bounds={bounds ?? COUNTRY_BOUNDS}
-        zoom={6}
-        maxZoom={10}
-        minZoom={6}
-        className="outline-none [&_.leaflet-interactive]:outline-none"
-        zoomControl={false}
+  return (
+    <MapContainer
+      bounds={bounds ?? COUNTRY_BOUNDS}
+      zoom={6}
+      maxZoom={10}
+      minZoom={6}
+      className="outline-none [&_.leaflet-interactive]:outline-none"
+      zoomControl={false}
+    >
+      <ZoomControls />
+
+      <MarkerClusterGroup
+        iconCreateFunction={clusterCustomIcon}
+        maxClusterRadius={100}
+        spiderfyOnMaxZoom={true}
+        showCoverageOnHover={true}
       >
-        <ZoomControls />
+        {initiatives.map((initiative) => (
+          <MapMarker
+            key={`marker_${initiative.initiativeId}`}
+            initiative={initiative}
+          />
+        ))}
+      </MarkerClusterGroup>
 
-        <MarkerClusterGroup
-          iconCreateFunction={clusterCustomIcon}
-          maxClusterRadius={100}
-          spiderfyOnMaxZoom={true}
-          showCoverageOnHover={true}
-        >
-          {initiatives.map((initiative) => (
-            <MapMarker
-              key={`marker_${initiative.initiativeId}`}
-              initiative={initiative}
-            />
-          ))}
-        </MarkerClusterGroup>
+      <ChangeView bounds={bounds ?? COUNTRY_BOUNDS} center={center} />
 
-        <ChangeView bounds={bounds ?? COUNTRY_BOUNDS} center={center} />
-
-        <GeoJSON
-          key={`geojson-layer-${processedData.length}`}
-          data={
-            {
-              type: "FeatureCollection",
-              features: processedData.map((d) => d.feature),
-            } as FeatureCollection<Geometry, DeptProperties>
-          }
-          style={setDeptStyle}
-          onEachFeature={setFeatureBehavior}
-        />
-
-        <TileLayer
-          key={`tile-layer-${tiles}`}
-          attribution={MAP_TILES[tiles].attribution}
-          url={MAP_TILES[tiles].url}
-        />
-      </MapContainer>
-
-      {/* NOTE: Al tener el componente en paralelo con el mapa se tiene acceso
-		  a lo que se calculó con este, sin el problema de bubbling que hay con
-		  elementos complejos dentro del mapa */}
-      <MapLegend
-        lowInitiativePerDepartment={min}
-        highInitiativePerDepartment={max}
-        departments={departmentsWithInitiatives}
-        tiles={tiles}
-        setTiles={setTiles}
-        layers={layer}
-        setLayers={setLayer}
+      <GeoJSON
+        key={`geojson-layer-${activeFeatures.length}`}
+        data={
+          {
+            type: "FeatureCollection",
+            features: activeFeatures.map((d) => d.feature),
+          } as FeatureCollection<Geometry, DeptProperties>
+        }
+        style={setDeptStyle}
+        onEachFeature={setFeatureBehavior}
       />
-    </>
+
+      <TileLayer
+        key={`tile-layer-${tiles}`}
+        attribution={MAP_TILES[tiles].attribution}
+        url={MAP_TILES[tiles].url}
+      />
+    </MapContainer>
   );
 }
